Fixup: Permit forwarding of ORD-containing blocks at tip if we are in IBD
Fixes a theoretical corner-case where if we are IBDing and some peer
asks us for the ord block and we have it at the tip, which is actively
being added-to, there is no reason why we shouldn't forward the block
along, since it clearly is not "new" since we are still building our
tip.

Not a huge deal either way but this `i` needed dotting.